refactor(node): move Job tests to sibling file per inline-tests guard
The guard freezes inline test-module debt shrink-only; the new tests in
tasks.rs were new debt. Same #[path] sibling pattern as error_tests.rs.
